[ALPS03571042] TEEI: kernel drivers
[Detail]
Modify the correct type for debug message.

MTK-Commit-Id: 92d451cc34887d511971d4d0a384219ef64ab1c5

CR-Id: ALPS03571042
Feature: Microtrust TEEI

+++ drivers/misc/mediatek/teei/V1.0/tz_driver/irq_register.c
@@ -143,7 +143,7 @@ static irqreturn_t nt_error_irq_handler(void)
 	unsigned long error_num = 0;
 
 	nt_get_secure_os_state((uint64_t *)(&error_num));
-	IMSG_ERROR("secure system ERROR ! error_num = %ld\n", (error_num - 4294967296));
+	IMSG_ERROR("secure system ERROR ! error_num = %ld\n", (unsigned long)(error_num - 4294967296));
 	soter_error_flag = 1;
 	up(&(boot_sema));
 	up(&smc_lock);
